\docType{package}
\name{biom-package}
\alias{biom-package}
\title{This is an R package for interfacing with the biom format.}
\description{
  This is an R package for interfacing with the
  biom-format. It includes basic utilities for reading and
  writing BIOM format using R, and native R methods for
  interacting with biom-data that maps to functionality in
  other languages that support biom-format, like python.
}
\author{
  Paul J. McMurdie II \email{mcmurdie@stanford.edu}
}
\references{
  \url{http://www.biom-format.org/}
}
\keyword{package}

