% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsBiom.R
\name{createClassif}
\alias{createClassif}
\title{A createClassif Function}
\usage{
createClassif(X.tr, Y.tr, learningSets, learnSetNames, selMethodNames,
  numGenes2Sel = c(3, 5, 10), classifierNames, cond3 = FALSE, isTunable,
  resultsDir, compName, numIter, ntoplist = 25, X.new, Y.new,
  validation = TRUE)
}
\arguments{
\item{X.tr}{matriz númerica con los valores de los datos. Donde las columnas son los "features" y las filas los individuos. Estos datos serán usados para la construcción del biomarcador}

\item{Y.tr}{vector de datos de clase "factor" donde se indican las condiciones experimentales (a comparar) de cada individuo indicado en las filas de X.tr}

\item{learningSets}{An object of class learningsets. May be missing, then the complete datasets is used as learning set.}

\item{learnSetNames}{método a utilizar para la generación de "learningsSets". Los diferentes métodos (LOOCV, CV, MCCV, bootstrap) se explican en "See Also"}

\item{selMethodNames}{vector caracter indicando métodos de seleccion (máximo 3?), ver posibles métodos en \code{\link{GeneSelection}} . (p.e: c("t.test"))}

\item{numGenes2Sel}{vector númeric donde se indican el número de genes incluidos en cada predictor.}

\item{classifierNames}{vector caracter indicando el método de classificador usado .(p.ej. c("CMA::dldaCMA", "CMA::rfCMA", "CMA::pnnCMA", "CMA::plrCMA") ), parámetro classifier \code{\link{GeneSelection}}}

\item{cond3}{valor lógico que indica si se la comparación, es decir, los factores son igual a 3.}

\item{resultsDir}{carpeta donde se guardaran los resultados. p.e. "results/". Solo necesario cuando saveLearnSet = TRUE}

\item{compName}{nombre de la comparación, se usara como identificador y para el nombre de archivos resultantes}

\item{numIter}{Number of iterations (s.details).}

\item{ntoplist}{número de features que se muestran en la lista de candidatos}

\item{X.new}{Solo si validation = TRUE. matriz númerica con los valores de los datos. Donde las columnas son los "features" y las filas los individuos. Estos datos serán usados para la validación del biomarcador}

\item{Y.new}{Solo si validation = TRUE. vector de datos de clase "factor" donde se indican las condiciones experimentales (a comparar) de cada individuo indicado en las filas de X.new}

\item{validation}{valor lógico que indica si se incluye una base de datos de validación.}

\item{fold}{Gives the number of CV-groups. Used only when method="CV"}
}
\description{
Construcción y validación externa de un biomarcador.
}
\examples{
# x <- cbind(matrix(rnorm(400,500),nrow = 40),matrix(rnorm(400,5),nrow = 40))
# colnames(x) <- paste0("a",1:ncol(x))
# rownames(x) <- paste0("aa",1:nrow(x))
# y <- factor(c(rep("A",20),rep("B",20)))
# lSet <- prepLearnSets(Y.tr = y , learnSetNames = "LOOCV", compName = "AvsB", saveLearnSet = F)  
# resF <- createClassif(X.tr = x,
#                       Y.tr = y,
  #                    learningSets  = lSet$learningSets,
    #                  learnSetNames = c("LOOCV"),
      #                selMethodNames = c("t.test"),
        #              numGenes2Sel = c(3,5,10),
          #            classifierNames = c("CMA::dldaCMA", "CMA::rfCMA", "CMA::pnnCMA") , 
            #          resultsDir = "results/",
              #        compName = "AvsB",
                #      validation=FALSE,
                  #    ntoplist = 10)
}
\author{
Miriam Mota <mmota.foix@gmail.com>
}
\seealso{
Text with \code{\link{GeneSelection}}
}
\keyword{biomarcador}
\keyword{clasificador}
\keyword{cma}
\keyword{learningsets}
\keyword{predictor}

