% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareComparison.R
\name{prepComp}
\alias{prepComp}
\title{A prepComp Function}
\usage{
prepComp(BD, cond, Y, appLog = FALSE)
}
\arguments{
\item{BD}{data.frame donde las filas son las muestras y las columnas los features además de una columna con el grupo o condición experimental}

\item{cond}{vector de strings que indique las diferentes condiciones experimentales en el orden que se desee. (p.e. c("bb","a"))}

\item{Y}{nombre de la columna en formato string donde se indica la condición experimental}

\item{appLog}{valor lógico que indica si se aplica logaritmo en base 2 a los datos.}
}
\value{
data.comp: data.frame original, donde las filas son las muestras (de las condiciones experimentales indicadas) y las columnas los "features". IMPORTANTE: no son necesariamente en formato númerico.

X: matriz numérica, donde las filas son las muestras (de las condiciones experimentales indicadas) y las columnas los "features"

Y: vector "factor", donde se indica la condición experimental correspondiente a cada una de las muestras
}
\description{
Prepara los datos de un comparación de m niveles para usar en las funciones prepLearnSets i createClassif.
}
\examples{
dat.prov <- data.frame(matrix(rnorm(1000),ncol=20))
colnames(dat.prov) <- paste0("gen",1:dim(dat.prov)[2])
rownames(dat.prov) <- paste0("sample",1:dim(dat.prov)[1])
dim(dat.prov)
dat.prov$condition <- c(rep("at.3",20),rep("b",20),rep("cc.4",10))
BvsA <- prepComp(BD = dat.prov, cond = c("b","a"), Y = "condition")
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{comparison}
\keyword{prepare}

