% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BIOM-class.R
\docType{methods}
\name{show,biom-method}
\alias{show}
\alias{show,biom-method}
\title{Method extensions to show for biom objects.}
\usage{
\S4method{show}{biom}(object)
}
\arguments{
\item{object}{biom-class object}
}
\description{
See the general documentation of \code{\link[methods]{show}} method for
expected behavior.
}
\examples{
# # # import with default parameters, specify a file
biom_file = system.file("extdata", "rich_sparse_otu_table.biom", package = "biomformat")
(x = read_biom(biom_file) )
show(x)
}
\seealso{
\code{\link[methods]{show}}
}

