% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allPackage.R
\docType{package}
\name{biom-package}
\alias{biom-package}
\title{This is an R package for interfacing with the biom format.}
\description{
This is an R package for interfacing with the biom-format.
It includes basic utilities for reading and writing BIOM format using R,
and native R methods for interacting with biom-data
that maps to functionality in other languages
that support biom-format, like python.
}
\author{
Paul J. McMurdie II \email{mcmurdie@stanford.edu}
}
\references{
\url{http://www.biom-format.org/}

 \url{https://github.com/joey711/biom}

 \url{http://cran.r-project.org/web/packages/biom/index.html}
}
\keyword{package}

